requirements.txt
python-dateutil>=2.8.1
requests>=2.25.1
tweepy>=3.10.0
psycopg2-binary>=2.9.1